Add dataset-wrangling code from this morning.
Ricky and I had a hacking session this morning to try to get D3 to load
our CSV dataset. Our dataset now loads, and we get back a huge array of
per-row JSON objects.

Is this inefficient? Yes, yes it is.
Does it work? Yes, yes it does.

Some example code is provided showing how to map functions across the
dataset, and how to use sets to extract only the unique values.

TODO:
 - Add some kind of animated spinner while the dataset loads.
   Materialize should have something for this.